Service e testes entidade Product implementados

+++ src/main/java/com/restaurant/restaurantManagement/model/dto/product/CreateProductDTO.java
@@ -29,7 +29,7 @@ public record CreateProductDTO(
 
         @NotNull
         @Positive
-        int minQuantityStock,
+        Long minQuantityStock,
 
         @NotNull
         InventoryDTO inventory
